Fix - Changes : sửa đổi thuộc tính của food , restaurant , order phần price , quantity về dạng BigDecimal , thêm model cho phần Coupones , sửa lại Database các thuộc tính đó , sửa các hàm có liên quan đến thuộc tính .

+++ src/main/java/DAO/CartDAO.java
@@ -244,13 +244,13 @@ public class CartDAO {
 				Food food = new Food();
 				food.setId(rs.getInt("FOOD_ID"));
 				food.setName(rs.getString("FNAME"));
-				food.setPrice(rs.getDouble("PRICE"));
+				food.setPrice(rs.getBigDecimal("PRICE"));
 				food.setImage(rs.getString("IMAGES"));
 				food.setResID(rs.getInt("RESID"));
 
 				CartItem item = new CartItem();
 				item.setDetailId(rs.getInt("DEATILID"));
-				item.setQuantity(rs.getInt("QUANTITY"));
+				item.setQuantity(rs.getBigDecimal("QUANTITY"));
 				item.setFood(food);
 
 				list.add(item);
